Fix assorted correctness bugs found via a code-review sweep
A collection of small, independent correctness fixes across the engine,
OS layer, and unit tests. Each is a localized defect (memory safety,
wrong operator/variable, resource leak, or ineffective test) with no
intended behavior change beyond the fix.

Engine (dev/ese/src):
- rec.cxx: delete[] for the array allocated in ErrRECIMakeKey; fix an &&
  that should be || in ErrIsamIPrereadIndexRanges (guards rgStartColumns[0]
  when cStartColumns == 0).
- space.cxx: null-check the PGNO array alloc and fix a std::sort off-by-one
  (cpgno-1 -> cpgno) in ErrSPCaptureSpaceTreePages.
- dbscan.cxx / dataserializer.cxx / inc/dataserializer.hxx: unique_ptr<T>
  owning new T[] -> unique_ptr<T[]>; add a missing null-check before memcpy.
- sysver.cxx: signed loop counter in ErrLGFindFmtVersFromEfv (a ULONG
  counter with i >= 0 never terminates and reads out of bounds).
- revertsnapshot.cxx / rbsdump.cxx: null-check several allocations, fix a
  leak on an ErrAddPage path, bound an on-disk length before use.
- _log/logwrite.cxx, _log/rstmap.cxx, inc/ccsr.hxx: == vs = in asserts; use
  the inner loop index in FRstmapCheckDuplicateSignature.
- inc/pib.hxx: ErrSetUserIoPriority assigns (=) instead of OR-accumulating
  the QOS priority so it can be lowered again.

Published header:
- published/inc/stat.hxx: off-by-one (> 1 -> > 0) in the monotonicity guard
  of both ErrGetSampleHits implementations.

OS layer / misc:
- os/string.cxx: advance past the NUL in the ErrOSSTRAsciiToUnicodeM
  multi-string loop (only the first substring was converted).
- os/norm.cxx: free the buffer on the GetLocaleInfo failure path.
- os/task.cxx: use the captured thread-node count during teardown.
- os/edbg.cxx: free szNewNote on the OOM cleanup path.
- os/blockcache/_hashedlrukcachechunk.hxx: operator-precedence fix.
- sync/sync.cxx: use the correct perf-data pointer in the term path.
- noncore/eseshadow/*.cxx: null-check WcsDupNew results; correct _countof
  buffer argument. noncore/interop/jetinterop.cpp: avoid leaking a heap
  MJET_TABLEID. _xpress9/Xpress9DecLz77.c: fix a misplaced SET_ERROR arg.

Tests:
- Correct wrong-variable/tautological assertions, a sizeof(pointer), an
  out-of-bounds read, a missing return on a detected inconsistency, a
  placement-new buffer mismatch, and a wildcard suffix-match fix.

+++ dev/ese/src/ese/sysver.cxx
@@ -193,7 +193,7 @@ ERR ErrGetDesiredVersion( _In_ const INST * const pinstStaging, _In_ JET_ENGINEF
         return JET_errSuccess;
     }
 
-    for ( ULONG i = g_ifmtversLastFeature; i >= 0; i-- )
+    for ( INT i = g_ifmtversLastFeature; i >= 0; i-- )
     {
         if ( g_rgfmtversEngine[i].efv == efvDesired )
         {
